control-plane/Python-src/requirements/base.txt
# Django core
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3

# Database
psycopg[binary]>=3.1

# Redis & Celery
redis>=5.0
celery>=5.3
django-celery-beat>=2.5

# WebSocket
channels>=4.0
channels-redis>=4.2
daphne>=4.1

# Authentication
djangorestframework-simplejwt>=5.3
PyJWT>=2.8

# YAML parsing
PyYAML>=6.0
jsonschema>=4.20

# Security
cryptography>=41.0

# Utilities
python-dotenv>=1.0
environs>=10.0
uuid6>=2023.5

# API documentation
drf-spectacular>=0.27
